Leave haltreq high while the target is halted.
This should help if a target unexpectedly resets while the debugger has
it halted. See #211.

I don't have a target to test this one. When I push reset on the Arty
board it also resets the Debug Module. (There's a small fix in this
change as well to prevent OpenOCD from increasing the busy delays when
that happens.)

Change-Id: I894bbadbfd1005fa74f5fbc1303697c6f513be13